Remove any \033[0m's from ls output ...
This assumes that all the ANSI colors that LS outputs affect foreground and it's
done so that I can set a background and it's going to still affect the whole
line.

We could replace all \033[0m's with \033[39m's to only reset foreground but with
LS, all ANSI colored stuff is at the end of the line.